Extract `APIResource`-related utils to a separate module
